mcpcrashcourse/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.6.0
groq==0.25.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema-pydantic==0.6
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.62
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.43
markdown-it-py==3.0.0
marshmallow==3.26.1
mcp==1.9.1
mcp-use==1.3.0
mdurl==0.1.2
multidict==6.4.4
mypy-extensions==1.1.0
numpy==2.2.6
orjson==3.10.18
packaging==24.2
propcache==0.3.1
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.9.1
pygments==2.19.1
python-dotenv==1.1.0
python-multipart==0.0.20
pyyaml==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.41
sse-starlette==2.3.5
starlette==0.46.2
tenacity==9.1.2
typer==0.16.0
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.1
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
yarl==1.20.0
zstandard==0.23.0
